Casi listo - Problemas con que cuenta debo pagar y con cual cobrar
