+++ src/utils/mercadoPago.js
@@ -1,0 +1,9 @@
+const mercadopago = require("mercadopago");
+
+mercadopago.configure({
+  access_token: process.env.ACCESS_TOKEN,
+});
+
+module.exports = {
+  mercadopago,
+};
